Make device log as dialog
