+++ src/views/Dashboard.svelte
@@ -1,7 +1,7 @@
 <script>
   import Navbar from '../components/Navbar.svelte';
 
-  import LogView from './LogView.svelte';
+  import DeviceLogDialog from './DeviceLogDialog.svelte';
   import RawJsonDialog from './RawJSONDialog.svelte';
 
   import NearbyDevicesTable from '../tables/NearbyDevicesTable.svelte';
@@ -11,9 +11,6 @@
   import { hash } from '../stores/hash';
 
   export let state;
-
-  let logLineWrap = false;
-  let logSearchText = '';
 </script>
 
 <Navbar />
@@ -44,27 +41,18 @@
     <h1>Incoming Connections</h1>
     <div class="connections-grid">
       <div class="connections-grid__left">
-      <!-- <div class="connections-grid__right"> -->
+        <!-- <div class="connections-grid__right"> -->
         <!-- <h2>Incoming</h2> -->
         <ConnectionsTable data={state.connectionsIn} />
       </div>
     </div>
   </section>
-
-  <!-- <section>
-    <div class="log__header">
-      <h1>Log</h1>
-      <div class="log__header__inputs">
-        <label> <input type="checkbox" bind:checked={logLineWrap} /> <span>Wrap lines</span> </label>
-        <div class="search-wrapper"><input type="search" bind:value={logSearchText} placeholder="Search log..." /></div>
-      </div>
-    </div>
-    <LogView data={state.log || []} lineWrap={logLineWrap} search={logSearchText} />
-  </section> -->
 </div>
 
-{#if $hash === '#json'}
-  <RawJsonDialog data={state} />
+{#if $hash === '#log'}
+  <DeviceLogDialog data={state.log || []} on:close={() => ($hash = '')} />
+{:else if $hash === '#json'}
+  <RawJsonDialog data={state} on:close={() => ($hash = '')} />
 {/if}
 
 <style>
@@ -94,27 +82,6 @@
   .connections-grid__right {
     flex-grow: 1;
     padding-top: 1rem;
-  }
-
-  .log__header {
-    display: flex;
-    justify-content: space-between;
-    align-items: center;
-    margin-bottom: 0.75rem;
-  }
-
-  .log__header h1 {
-    margin: 0;
-  }
-
-  .log__header__inputs {
-    display: flex;
-    justify-content: flex-end;
-    align-items: center;
-  }
-
-  .search-wrapper {
-    margin-left: 1.5rem;
   }
 
   @media (min-width: 1024px) {
